+++ tests/home-seo-signals.test.ts
@@ -3,6 +3,7 @@ import { readFileSync } from 'node:fs';
 import test from 'node:test';
 
 const homeSource = readFileSync('frontend/app/(localized)/[locale]/(marketing)/(home)/page.tsx', 'utf8');
+const homeJsonLdSource = readFileSync('frontend/app/(localized)/[locale]/(marketing)/(home)/_lib/home-jsonld.ts', 'utf8');
 const homeSectionsSource = [
   readFileSync('frontend/components/marketing/home/HomeRedesignSections.tsx', 'utf8'),
   readFileSync('frontend/components/marketing/home/HomeConversionSections.tsx', 'utf8'),
@@ -46,8 +47,9 @@ test('homepage titles keep the pay-as-you-go differentiator without getting too 
 test('homepage structured data keeps Organization schema alongside WebApplication schema', () => {
   assert.match(homeSource, /home-webapp-jsonld/);
   assert.match(homeSource, /home-organization-jsonld/);
-  assert.match(homeSource, /'@type': 'Organization'/);
-  assert.match(homeSource, /logo: 'https:\/\/maxvideoai\.com\/favicon-512\.png'/);
+  assert.match(homeSource, /buildOrganizationSchema\(\)/);
+  assert.match(homeJsonLdSource, /'@type': 'Organization'/);
+  assert.match(homeJsonLdSource, /logo: 'https:\/\/maxvideoai\.com\/favicon-512\.png'/);
 });
 
 test('homepage FAQ targets search-intent questions and shares the same items with FAQPage schema', () => {
